Added a router to the deckbuilder to facilitate multiple pages, better switching between components, better performance and better maintenence.
