+++ package.json
@@ -8,6 +8,7 @@
     "react-dom": "^16.9.0",
     "react-ga": "^2.6.0",
     "react-list-drag-and-drop": "^0.9.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.1.1"
   },
   "scripts": {
