+++ src/App.js
@@ -1,20 +1,31 @@
-import './App.css'; 
 import React            from 'react';
-import DeckBuilderForm  from './libs/DeckBuilderForm.js';
+import AppLayout        from './layout/AppLayout.js';
 import GoogleHelper     from './libs/GoogleHelper.js';
-import logo             from './logo.svg'
-import AnalyticsHelper  from './libs/AnalyticsHelper.js';
-import {
-  Layout,     Menu, 
-  Typography, Steps, 
-  Button,     message, 
-  Icon,       Spin,
-  Result
-}                       from 'antd';
 
-const {Header, Content} = Layout;
-const {Title, Text}     = Typography;
-const {Step}            = Steps;
+import {
+  Typography, 
+  message,
+  Icon,
+  Button,
+  Result
+} from 'antd';
+
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Redirect
+} from "react-router-dom";
+
+// Pages
+import ChooseFolder from './layout/pages/ChooseFolder.js';
+import DeckBuilder  from './layout/pages/DeckBuilder.js';
+import SuccessPage  from './layout/pages/SuccessPage.js';
+
+// CSS
+import './App.css'; 
+
+const {Text} = Typography;
 
 class App extends React.Component {
 
@@ -28,13 +39,11 @@ class App extends React.Component {
 
     // Set the initial state of the component
     this.state = {
+      googleLoaded:       false,  // Whether the Google SDK has loaded yet
       loggedIn:           null,   // Whether the user is successfully logged in to Google or not
       team:               'se',   // This could be used in the future to build other team's decks
       destinationFolder:  null,   // This is the folder id for the chosen Drive folder
       googleUsername:     null,   // This is the user's name for the welcome message
-      current:            0,      // Current step - start at 0
-      generating:         false,  // Are we currently in the process of generating a deck?
-      generatingMessage:  "",     // The message we will show while we are generating the deck (will change over time)
       deckUrl:            null,   // The ID of the new presentation so we can link to it at the end
       notifsAllowed:      false,  // Whether notifications are allowed based on browser permissions
       shouldNotify:       true    // Whether the user wants us to notify them or not
@@ -42,9 +51,6 @@ class App extends React.Component {
 
     // Load the Google Helper for use in this component
     this.googleHelper = new GoogleHelper();
-
-    // Load the Analytics Helper for use in this component
-    this.analyticsHelper = new AnalyticsHelper();
   }
 
   /**
@@ -58,7 +64,7 @@ class App extends React.Component {
       // Check if they user is already signed in
       if(this.googleHelper.isSignedIn) {
         // If they are logged in then we can go ahead and proceed
-        this.handleGoogleLogin();
+        this.afterLogin();
       }
       else {
         // If they aren't logged in we will update the state here
@@ -68,8 +74,6 @@ class App extends React.Component {
         });
       }
     });
-
-    this.handleNotifications();
   }
 
   /**
@@ -79,130 +83,65 @@ class App extends React.Component {
   render() {
     // Set up the current state of the component for the rendering below
     // This is just to make the following code more manageable
-    const current         = this.state.current,
-          loggedIn        = this.state.loggedIn,
+    const loggedIn        = this.state.loggedIn,
           googleUsername  = this.state.googleUsername,
-          folderName      = (this.state.destinationFolder ? this.state.destinationFolder.name : null),
+          folder          = this.state.destinationFolder,
+          folderName      = folder !== null ? folder.name : "",
           deckUrl         = this.state.deckUrl,
           team            = this.state.team,
           notifsAllowed   = this.state.notifsAllowed,
           shouldNotify    = this.state.shouldNotify;
 
-    // TODO: Split this out into separate components for readability
     return (
-      <Layout className="app">
-        <Header>
-          <img alt="" className="logo" src={logo} style={{height: 31, margin: '16px 24px 16px 0', float: 'left'}} />
-          <Menu
-            theme="dark"
-            mode="horizontal"
-            defaultSelectedKeys={['1']}
-            style={{ lineHeight: '64px' }}
-          >
-            <Menu.Item key="1">DeckBuilder</Menu.Item>
-          </Menu>
-        </Header>
-        <Content style={{ padding: '50px' }}>
-          <div style={{ background: '#fff', padding: 24 }}>
-
-            <Title 
-              level={2}
-              style={{marginBottom: 25}}
-            >Optimizely SE Deck Builder</Title>
-
-            <Steps size="small" current={current} style={{marginBottom: 25}}>
-              <Step title="Login to Google" description={loggedIn ? "Hi, " + googleUsername : ""} />
-              <Step title="Choose folder" description={folderName ? folderName : ""} />
-              <Step title="Configure Slides" />
-            </Steps>
-            
-            {/* Page for logging into Google */}
-            {current === 0 && (
-              <React.Fragment>
-                {loggedIn === null && (
-                  <Text type="secondary">
-                    <Icon type="loading" /> &nbsp;
-                    Checking Google login status
-                  </Text>
-                )}
-
-                {loggedIn === false && (
-                  <div className="steps-action" style={{marginTop: 25}}>
-                    <Button type="primary" onClick={() => this.signInGoogle()}>
-                      Sign into Google
-                    </Button>
-                  </div>
-                )}
-              </React.Fragment>
-            )}
-
-            {/* Page for choosing your folder */}
-            {current === 1 && (
-              <Text type="secondary">
-                <Icon type="loading" /> &nbsp;
-                Selecting folder
-              </Text>
-            )}
-
-            {/* Page for configuring your slide deck */}
-            {current === 2 && (
-              <Spin tip={this.state.generatingMessage} spinning={this.state.generating}>
-                <DeckBuilderForm 
+      <Router>
+        <AppLayout
+          username={googleUsername}
+          folderName={folder ? folder.name : null} 
+        >
+          {loggedIn ? (
+            <Switch>
+              <Route exact path="/" render={() => (
+                  <Redirect to="/choose-folder"/>
+              )}/>
+              <Route path="/choose-folder">
+                <ChooseFolder 
+                  googleHelper={this.googleHelper} 
+                  folderChosen={(folder) => this.handleFolderChosen(folder)}
+                />
+              </Route>
+              <Route path="/deck-builder">
+                <DeckBuilder 
                   team={team} 
-                  generate={(values, chosenDecks, deletedDecks) => this.generateDeck(values, chosenDecks, deletedDecks)}
-                  notificationsAllowed={notifsAllowed}
-                  shouldNotify={shouldNotify}
-                  notifyCallback={(shouldNotify) => this.shouldNotifyChanged(shouldNotify)}
+                  googleHelper={this.googleHelper}
+                  folder={folder}
                   seName={googleUsername}
                 />
-              </Spin>
-            )}
+              </Route>
+              <Route path="/success">
+                <SuccessPage folderName={folderName} />
+              </Route>
+            </Switch>
+          ) : (
+            <React.Fragment>
+              {loggedIn === null && (
+                <Text type="secondary">
+                  <Icon type="loading" /> &nbsp;
+                  Checking Google login status
+                </Text>
+              )}
 
-            {/* Page for showing the result! */}
-            {current === 3 && (
-              <Result
-                status="success"
-                title="Your presentation is ready!"
-                subTitle={"We have successfully configured your slide deck. You can find it in your Google Drive folder "+folderName+" or access it by clicking below:"}
-                extra={[
-                  <Button type="primary" href={deckUrl} target="_blank" key="open">
-                    Open presentation
+              {loggedIn === false && (
+                <div className="steps-action" style={{marginTop: 25}}>
+                  <Button type="primary" onClick={() => this.handleLogin()}>
+                    Sign into Google
                   </Button>
-                ]}
-              />
-            )}
-          </div>
-        </Content>
-      </Layout>
+                </div>
+              )}
+            </React.Fragment>
+          )}
+        </AppLayout>
+      </Router>
     );
-  }
-
-  ///////////////////
-  // Notifications //
-  ///////////////////
-
-  handleNotifications() {
-    let permission = Notification.permission;
-
-    if(permission === "default") {
-      // Ask the customer to allow notifications
-      Notification.requestPermission((permission) => {
-        this.setState({
-          notifsAllowed: (permission === "granted")
-        })
-      });
-    }
-    else {
-      this.setState({
-        notifsAllowed: (permission === "granted")
-      })
-    }
-  }
-
-  shouldNotifyChanged(shouldNotify) {
-    this.setState({
-      shouldNotify: shouldNotify
-    });
   }
 
   ///////////////////////////////////////
@@ -210,272 +149,31 @@ class App extends React.Component {
   ///////////////////////////////////////
 
   /**
-   * Handle button click to sign into google
-   * @return {null}
-   */
-  signInGoogle() {
-    this.googleHelper.signIn()
-      .then(() => {
-        this.handleGoogleLogin();
-      });
-  }
-
-  /**
    * Once the user is logged into Google, set the state and progress to the next stage
    * @return {null}
    */
-  handleGoogleLogin() {
+  handleLogin() {
+    this.googleHelper.signIn()
+    .then(() => {
+      this.afterLogin();
+    });
+  }
+
+  afterLogin() {
     // Get the google username and increment the step
     this.setState({
       googleUsername: this.googleHelper.getUsername(),
-      current: this.state.current + 1,
       loggedIn: true
     });
 
     // Send a visual message to tell user they're logged in
     message.success('Logged in as ' + this.state.googleUsername);
-
-    // Run the next step (choosing a folder)
-    this.googleHelper.createPicker((data) => this.pickerCallback(data));
   }
 
-  pickerCallback(data) {
-    if (data.action === window.google.picker.Action.PICKED) {
-      // Add foldername to the App state and increment stage
-      this.setState({
-        destinationFolder: data.docs[0],
-        current: this.state.current + 1
-      });
-    }
-  }
-
-  ///////////////////////////////
-  // Deck generating functions //
-  ///////////////////////////////
-
-  /**
-   * Handle the "generate" button click and start processing the new
-   *   presentation with google
-   * @param  {object} values        The values returned by the form
-   * @param  {array} originalDecks  The original list of decks from the
-   *                                deckbuilder
-   * @return {null}               
-   */
-  generateDeck(values, chosenDecks, deletedDecks) {
-    // Set state so that we can show the loading spinner
+  handleFolderChosen(folder) {
     this.setState({
-      generating: true,
-      generatingMessage: "Copying master deck to new location"
+      destinationFolder: folder
     });
-
-    this.analyticsHelper.trackState("generate clicked");
-
-    // Track decks via analytics
-    for(const deck of chosenDecks) {
-      this.analyticsHelper.track(deck);
-    }
-
-    // Get relevant variables from the values object
-    const customerName  = values.customer_name;
-
-    // Generate deck filename
-    const filename = this.generateFilename(customerName);
-
-    // Copy the master deck into the destination folder
-    this.googleHelper.copyMasterDeck(filename, this.state.destinationFolder.id)
-    .then((fileId) => {
-
-      // Update the loading text on the spinner and the file ic
-      this.setState({
-        generatingMessage: "Accessing new presentation",
-        deckUrl: "https://docs.google.com/presentation/d/" + fileId
-      });
-
-      // Get the presentation slides
-      this.googleHelper.getPresentation(fileId)
-      .then((masterDeck) => {
-
-        // Update the loading text on the spinner
-        this.setState({
-          generatingMessage: "Configuring slides (this may take a while)"
-        });
-
-        // Get the slides out of this presentation
-        const slides = masterDeck.slides;
-
-        // Update the title slide with text from our DeckBuilder
-        this.updateTitleAndAgendaSlides(fileId, slides, chosenDecks, values);
-
-        // Add the customer logo
-        this.addCustomerLogoToDeck(fileId, values.logo, slides);
-
-        // Check if we need to delete any slides
-        if(deletedDecks.length > 0) {
-          // Delete unnecessary slides
-          this.deleteSlides(fileId, slides, deletedDecks)
-          .then(() => { 
-            // Call to finish
-            this.finishedGenerating();
-          });
-        }
-        else {
-          // No slides need to be deleted, we can just finish
-          this.finishedGenerating();
-        }
-
-      });
-
-    });
-  }
-
-  /**
-   * Generate a filename for the new deck
-   * @param  {string} prepend The text to be prepended onto the filename
-   * @return {string}         The new filename
-   */
-  generateFilename(prepend) {
-    return prepend + " | Optimizely Overview";
-  }
-
-  /**
-   * Update the title and agenda slides in the new presentation deck
-   * @param  {string} fileId        The presentation's file id
-   * @param  {array}  slides        A list of the slides in the presentation
-   *                                deck
-   * @param  {object} values        Values from the deckbuilder form that was
-   *                                submitted
-   * @param  {array}  originalDecks The original decks from our config file
-   * @return {Promise}              Gets returned when the updates are made      
-   */
-  updateTitleAndAgendaSlides(fileId, slides, chosenDecks, values) {
-    // Generate the agenda text
-    let agendaInt   = 1;
-    let agendaText  = "";
-    for(const deck of chosenDecks) {
-      agendaText += agendaInt + ". " + deck.agendaTitle + "\n";
-      agendaInt++;
-    }
-
-    // Generate the replacements we want to implement
-    const replacements = [
-      {
-        slide: slides[0],
-        searchString: "COMPANYNAME",
-        replacementString: values.customer_name
-      },
-      {
-        slide: slides[0],
-        searchString: "AEName",
-        replacementString: values.ae_name
-      },
-      {
-        slide: slides[0],
-        searchString: "SENAME",
-        replacementString: values.se_name
-      },
-      {
-        slide: slides[1],
-        searchString: "AGENDAHERE",
-        replacementString: agendaText
-      }
-    ];
-
-    // Perform the replacements
-    this.googleHelper.updateSlidesByReplacingText(fileId, replacements);
-  }
-
-  addCustomerLogoToDeck(deckId, logoURL, slides) {
-    // Find the correct slide
-    let slide = slides[0];
-
-    // Add it
-    this.googleHelper.addLogoToSlide(deckId, logoURL, slide);
-  }
-
-  /**
-   * Delete the slides that weren't chosen by the user
-   * @param  {string}  fileId        The ID of the presentation deck that has been generated
-   * @param  {array}   originalDecks The array of decks that we pulled from config
-   * @param  {array}   chosenDecks   The list of decks that the user has chosen
-   * @return {Promise}               A promise when the requests complete
-   */
-  deleteSlides(fileId, slides, deletedDecks) {
-    return new Promise((resolve, reject) => {
-      // Get the list of deletions - this can go straight into the request
-      const deletions = this.generateDeletions(deletedDecks, slides)
-
-      // Delete the decks
-      this.googleHelper.deleteSlides(fileId, deletions)
-      .then(() => resolve());
-    });
-  }
-
-  /**
-   * Generate array of slides to be deleted
-   * @param  {array} originalDecks The original list of decks that have the
-   *   offsets and slide lengths in them
-   * @param  {array} chosenDecks   The list of decks that have been chosen by
-   *   the user
-   * @param  {array} slides        The slides from the new deck we retrieved
-   *   from the Google API (this will have all of the slide ids in it for us
-   *   to record)
-   * @return {array}               Array of objects which will form the
-   *   request to batch delete slides from the presentation 
-   */
-  generateDeletions(deletedDecks, slides) {
-    // Create variable to store list of deletions
-    let deletions = [];
-
-    // Loop through all the original decks to see which ones we need to delete.
-    for(const deckToDelete of deletedDecks) {
-      // Start at the current deck's offset in the master deck 
-      let currOffset = deckToDelete.offset;
-
-      // Loop through each subsequent slide
-      for(var j = 0; j < deckToDelete.slides; j++) {
-        // Get the slide
-        let slide = slides[currOffset];
-
-        // Get the slide object id
-        let objectId = slide.objectId;
-
-        // Add it to be deleted
-        deletions.push(objectId);
-
-        // Increment the offset
-        currOffset++;
-      }
-    }
-    return deletions;
-  }
-
-  /**
-   * Called once the presentation deck has finished generating
-   * @return {null} 
-   */
-  finishedGenerating() {
-    // Move the state on to the final screen
-    this.setState({
-      current: this.state.current + 1
-    });
-
-    this.analyticsHelper.trackState("finished generating");
-
-
-    // Check if the user wants to be notified
-    if(this.state.shouldNotify) {
-      // Send the notification
-      let notif = new Notification("Optimizely Deck Builder", {
-        "body": "Your deck is ready; click here to open",
-        "icon": "https://upload.wikimedia.org/wikipedia/en/thumb/e/e9/Optimizely_Logo.png/220px-Optimizely_Logo.png"
-      });
-
-      // Override click function
-      notif.onClick = function(event) {
-        event.preventDefault(); // prevent the browser from focusing the Notification's tab
-        window.open(this.state.deckUrl, "_blank");
-      }
-    }
   }
 }
